Move about and education pages to components folder
